src/vataproductivity/requirements.txt
arrow==1.1.1
binaryornot==0.4.4
briefcase==0.3.5
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
importlib-metadata==4.6.3
Jinja2==2.11.3
jinja2-time==0.2.0
MarkupSafe==2.0.1
poyo==0.5.0
pycparser==2.20
python-dateutil==2.8.2
python-slugify==5.0.2
pythonnet==2.5.2
requests==2.26.0
six==1.16.0
smmap==4.0.0
text-unidecode==1.3
toga-core==0.3.0.dev28
toga-winforms==0.3.0.dev28
toml==0.10.2
travertino==0.1.3
typing-extensions==3.10.0.0
urllib3==1.26.6
